third_party/WebKit/LayoutTests/svg/custom/path-segment-initial-role.html

<!DOCTYPE html>
<html>
  <body>
    <script src="../../resources/js-test.js"></script>
    <script>
      var pathElement = document.createElementNS("http://www.w3.org/2000/svg","path");

      var pathSegment = pathElement.createSVGPathSegCurvetoCubicAbs(0.0, 1.0, 2.0, 3.0, 4.0, 5.0);
      pathSegment.x1 = -1.0;

      shouldBe('pathSegment.x1', '-1.0');
    </script>
	<p id="description">Test that there is no assertion when setting values on a new path segment not in any segment list.</p>
	<div id="console"></div>
  </body>
</html>
